update to stable

+++ wifi802_11.cpp
@@ -1,4 +1,4 @@
-#include"wifi802_11.h"
+#include "wifi802_11.h"
 #ifdef ESP32
 #include <WiFi.h>
 #include "esp_wifi.h"
@@ -8,7 +8,7 @@
 #include <ESP8266WiFi.h>
 #include <user_interface.h>
 #endif
-#include<Arduino.h>
+#include <Arduino.h>
 
 const char *ssid = "MESH_NETWORK";
 char wifi_password[20];
@@ -21,19 +21,20 @@ char wifi_password[20];
 #define DATA_START_OFFSET 24
 
 uint8_t raw_HEADER[] = {
-  //MAC HEADER
-  0x40, 0x0C,             // 0-1: Frame Control  //Version 0 && Data Frame && MESH
-  0x00, 0x00,             // 2-3: Duration
-  0xaa, 0xbb, 0xcc, 0xee, 0xff, 0x11,       // 4-9: Destination address (broadcast)
-  0xba, 0xde, 0xaf, 0xfe, 0x00, 0x06,       // 10-15: Source address
-  0xba, 0xde, 0xaf, 0xfe, 0x00, 0x06,       // 16-21: BSSID
-  0x00, 0x00             // 22-23: Sequence / fragment number
+    //MAC HEADER
+    0x40, 0x0C,                         // 0-1: Frame Control  //Version 0 && Data Frame && MESH
+    0x00, 0x00,                         // 2-3: Duration
+    0xaa, 0xbb, 0xcc, 0xee, 0xff, 0x11, // 4-9: Destination address (broadcast)
+    0xba, 0xde, 0xaf, 0xfe, 0x00, 0x06, // 10-15: Source address
+    0xba, 0xde, 0xaf, 0xfe, 0x00, 0x06, // 16-21: BSSID
+    0x00, 0x00                          // 22-23: Sequence / fragment number
 };
 short sequence = 0;
-void(*wifi_802_receive_callback)(const uint8_t *, int, uint8_t) = NULL;
+void (*wifi_802_receive_callback)(const uint8_t *, int, uint8_t) = NULL;
 
 #ifdef ESP32
-void receive_raw_cb(void *recv_buf, wifi_promiscuous_pkt_type_t type) {
+void receive_raw_cb(void *recv_buf, wifi_promiscuous_pkt_type_t type)
+{
   /*
   20:35:41.710 ->            BF 20 24 80 00 00 00 00 A1 00 01 B6 AC 32 16 00 __$__________2__
   20:35:41.710 ->            93 0A 06 22 00 00 60 63 24 40 02 00 40 0C 00 00 ___"__`c$@__@___
@@ -46,18 +47,21 @@ void receive_raw_cb(void *recv_buf, wifi_promiscuous_pkt_type_t type) {
 
   */
   wifi_promiscuous_pkt_t *sniffer = (wifi_promiscuous_pkt_t *)recv_buf;
-  if(sniffer->payload[0]!=0x40) return;
-  if(memcmp(sniffer->payload+BSSID_OFFSET,raw_HEADER+BSSID_OFFSET, 6)!=0) return;
+  if (sniffer->payload[0] != 0x40)
+    return;
+  if (memcmp(sniffer->payload + BSSID_OFFSET, raw_HEADER + BSSID_OFFSET, 6) != 0)
+    return;
 
-  unsigned char *d = sniffer->payload+DATA_START_OFFSET;
-  short length = ((unsigned short)d[0])<<8 | d[1];
+  unsigned char *d = sniffer->payload + DATA_START_OFFSET;
+  short length = ((unsigned short)d[0]) << 8 | d[1];
 
-  wifi_802_receive_callback(d+2, length,sniffer->rx_ctrl.rssi);
+  wifi_802_receive_callback(d + 2, length, sniffer->rx_ctrl.rssi);
 
   return;
 }
 #else
-void receive_raw_cb(unsigned char*frm, short unsigned int len) {
+void receive_raw_cb(unsigned char *frm, short unsigned int len)
+{
   /*
     16:34:05.795 ->            C3 10 26 50 00 00 00 00 00 00 01 00 [40 0C 00 00 __&P________@___
     16:34:05.795 ->            FF FF FF FF FF FF BA DE AF FE 00 [06] BA DE AF FE ________________
@@ -71,61 +75,68 @@ void receive_raw_cb(unsigned char*frm, short unsigned int len) {
   uint8_t rssi = frm[0];
   //if(frm[0]!=0x40) return;
 
-  if(frm[12]!=0x40) return;
-  if(memcmp(frm+BSSID_OFFSET+12,raw_HEADER+BSSID_OFFSET, 6)!=0) return;
-  unsigned char *d = frm+12+DATA_START_OFFSET;
+  if (frm[12] != 0x40)
+    return;
+  if (memcmp(frm + BSSID_OFFSET + 12, raw_HEADER + BSSID_OFFSET, 6) != 0)
+    return;
+  unsigned char *d = frm + 12 + DATA_START_OFFSET;
 
-  short length = ((unsigned short)d[0])<<8 | d[1];
+  short length = ((unsigned short)d[0]) << 8 | d[1];
 
-  if(wifi_802_receive_callback!=NULL) {
-    wifi_802_receive_callback(d+2, length, rssi);
+  if (wifi_802_receive_callback != NULL)
+  {
+    wifi_802_receive_callback(d + 2, length, rssi);
   }
 }
 #endif
 char password[15];
-void wifi_802_11_begin(char bsId[], int channel){
+void wifi_802_11_begin(char bsId[], int channel)
+{
   //WiFi.begin();
   String mac = WiFi.macAddress();
-  memcpy(raw_HEADER+BSSID_OFFSET, bsId, 6);
-  memcpy(raw_HEADER+MY_MAC_OFFSET, mac.c_str(), 6);
+  memcpy(raw_HEADER + BSSID_OFFSET, bsId, 6);
+  memcpy(raw_HEADER + MY_MAC_OFFSET, mac.c_str(), 6);
 
-  #ifdef ESP32
+#ifdef ESP32
   esp_wifi_set_mode(WIFI_MODE_STA);
   esp_wifi_set_channel(channel, WIFI_SECOND_CHAN_NONE);
   esp_wifi_set_promiscuous_rx_cb(receive_raw_cb);
   esp_wifi_set_promiscuous(1);
   esp_wifi_set_max_tx_power(127);
-  #else
+#else
   wifi_set_opmode(STATION_MODE);
   wifi_set_channel(channel);
   wifi_set_promiscuous_rx_cb(receive_raw_cb);
   wifi_promiscuous_enable(true);
   WiFi.setOutputPower(20.5);
-  #endif
+#endif
 }
 
-void wifi_802_receive_cb(void(*cb)(const uint8_t *, int, uint8_t)) {
-    wifi_802_receive_callback = cb;
+void wifi_802_receive_cb(void (*cb)(const uint8_t *, int, uint8_t))
+{
+  wifi_802_receive_callback = cb;
 }
 
-void wifi_802_11_send(const uint8_t *d, int len) {
+void wifi_802_11_send(const uint8_t *d, uint16_t len)
+{
   uint8_t buf[500];
-  for (int i=0;i<5;i++) { // really ?
-    if (len>sizeof(buf)-sizeof(raw_HEADER)-2) return;
+  for (int i = 0; i < 5; i++)
+  { // really 5 times ?
+    if (len > sizeof(buf) - sizeof(raw_HEADER) - 2)
+      return;
 
-    memcpy(buf,raw_HEADER, sizeof(raw_HEADER));
-    memcpy(buf+sizeof(raw_HEADER)+2, d, len);
-    memcpy(buf+SEQNUM_OFFSET,(char*)&sequence, 2);
+    memcpy(buf, raw_HEADER, sizeof(raw_HEADER));
+    memcpy(buf + sizeof(raw_HEADER) + 2, d, len);
+    memcpy(buf + SEQNUM_OFFSET, (char *)&sequence, 2);
 
-    buf[sizeof(raw_HEADER)]=(len>>8)&0xff;
-    buf[sizeof(raw_HEADER)+1]=len&0xff;
+    buf[sizeof(raw_HEADER)] = (len >> 8) & 0xff;
+    buf[sizeof(raw_HEADER) + 1] = len & 0xff;
 
-
-      #ifdef ESP32
-      esp_wifi_80211_tx(ESP_IF_WIFI_STA, buf, sizeof(raw_HEADER) + len+ 2, true);
-      #else
-      wifi_send_pkt_freedom(buf, sizeof(raw_HEADER) + len+ 2, true);
-      #endif
-      sequence++;
+#ifdef ESP32
+    esp_wifi_80211_tx(ESP_IF_WIFI_STA, buf, sizeof(raw_HEADER) + len + 2, true);
+#else
+    wifi_send_pkt_freedom(buf, sizeof(raw_HEADER) + len + 2, true);
+#endif
+    sequence++;
   }
 }
