+++ wifi802_11.h
@@ -3,7 +3,7 @@
 #include <Arduino.h>
 
 void wifi_802_11_begin(char bsId[], int channel);
-void wifi_802_11_send(const uint8_t *d, int len);
+void wifi_802_11_send(const uint8_t *d, uint16_t len);
 void wifi_802_receive_cb(void(*cb)(const uint8_t *, int, uint8_t));
 
 #endif
